Add interface for clock
