+++ src/Shared/Metrics/Manager/MemoryMetricManager.php
@@ -7,7 +7,7 @@ use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\Metric;
 use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\MetricManager;
 use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\MetricSubject;
 use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\Record;
-use ArtARTs36\MergeRequestLinter\Shared\Time\Clock;
+use ArtARTs36\MergeRequestLinter\Shared\Time\LocalClock;
 use Psr\Clock\ClockInterface;
 
 class MemoryMetricManager implements MetricManager
